pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = [ "setuptools", "wheel" ]

[project]
name = "romrescue"
version = "0.0.1"
description = "Website for SOS Romanian Rescue"
authors = [ { name = "Ben Gosney", email = "bengosney@googlemail.com" } ]
requires-python = ">=3.10"
classifiers = [
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]
dependencies = [ "wagtail" ]

optional-dependencies.dev = [
  "django-debug-toolbar",
  "django-stubs",
  "icecream",
  "mypy",
  "pip-tools",
  "pre-commit",
  "ruff",
]

[tool.setuptools]
license-files = [ "LICENSE" ]

[tool.black]
line-length = 128
target-version = [ 'py312' ]

[tool.ruff]
target-version = "py312"
line-length = 128
exclude = [
  "migrations",
]

lint.select = [ "C90", "DJ", "E4", "E7", "E9", "F", "I", "N", "UP" ]

lint.isort.combine-as-imports = true
lint.isort.split-on-trailing-comma = false

[tool.pycln]
all = true

[tool.mypy]
plugins = [ "mypy_django_plugin.main" ]

[tool.django-stubs]
django_settings_module = "stl.settings"

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "pep440"
version_provider = "pep621"
update_changelog_on_bump = true
